requirements.txt
tsfresh==0.16.0
sqlalchemy==1.3.19
mysql==0.0.2
pymars==0.5.1
pandas>=1.0.0
protobuf==3.15.8